+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 ### Changed
-- Capture Error returns optional
+- Capture Error returns optional and renamed
 ### Removed
 
 ## [1.4.0] - 2023-03-12